requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
clickhouse-connect==0.6.23
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
duckdb==0.9.2
effdet==0.4.1
emoji==2.9.0
fastapi==0.108.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.340
langchain-community==0.0.7
langchain-core==0.1.4
langdetect==1.0.9
langsmith==0.0.75
layoutparser==0.3.4
lxml==5.0.0
lz4==4.3.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.6.1
opencv-python==4.9.0.80
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-anthropic==0.6.0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-bedrock==0.6.0
opentelemetry-instrumentation-chromadb==0.6.0
opentelemetry-instrumentation-cohere==0.6.0
opentelemetry-instrumentation-dbapi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-langchain==0.6.0
opentelemetry-instrumentation-llamaindex==0.6.0
opentelemetry-instrumentation-openai==0.6.0
opentelemetry-instrumentation-pinecone==0.6.0
opentelemetry-instrumentation-pymysql==0.42b0
opentelemetry-instrumentation-replicate==0.7.0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-transformers==0.6.0
opentelemetry-instrumentation-urllib3==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-semantic-conventions-ai==0.0.12
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.11.2
pillow==10.2.0
portalocker==2.8.2
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.1
pydantic==1.10.7
pydantic_core==2.14.6
pyparsing==3.1.1
pypdf==3.17.4
pypdfium2==4.25.0
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
starlette==0.32.0.post1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traceloop-sdk==0.6.0
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.11.6
unstructured-client==0.15.1
unstructured-inference==0.7.20
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
